refactor(vscode): extract SessionStreamScheduler and expand coverage
Move the scheduler into its own file under src/kilo-provider/ alongside the other provider helpers. Background lane now recomputes its flush target on every push so adaptive backoff kicks in within a single streaming burst, bounded by backgroundMaxMs. Adds a dedicated test file covering coalescing, session isolation, focus/lifecycle, adaptive scheduling, backoff cap, and stats.
