+++ packages/kilo-vscode/tests/unit/kilo-provider-utils.test.ts
@@ -9,7 +9,6 @@ import {
   mapCloudSessionMessageToWebviewMessage,
   MessageConfirmation,
   mergeFileSearchResults,
-  SessionStreamScheduler,
   type ProviderInfo,
 } from "../../src/kilo-provider-utils"
 import type { CloudSessionMessage } from "../../src/services/cli-backend/types"
@@ -474,250 +473,6 @@ describe("mapSSEEventToWebviewMessage", () => {
   it("returns null for unhandled event types (Like global.disposed)", () => {
     const event: Event = { type: "global.disposed", properties: {} }
     expect(mapSSEEventToWebviewMessage(event, undefined)).toBeNull()
-  })
-})
-
-describe("session stream scheduling", () => {
-  function update(text: string, delta?: string, sid = "sess-1", partID = "p1") {
-    const msg = {
-      type: "partUpdated" as const,
-      sessionID: sid,
-      messageID: "m1",
-      part: { id: partID, type: "text", messageID: "m1", text },
-    }
-    if (delta === undefined) return msg
-    return { ...msg, delta: { type: "text-delta" as const, textDelta: delta } }
-  }
-
-  type Item = ReturnType<typeof update>
-  type Sent = Item | { type: "partsUpdated"; updates: Item[] }
-
-  function items(sent: Sent[]) {
-    return sent.flatMap((msg) => (msg.type === "partsUpdated" ? msg.updates : [msg]))
-  }
-
-  function flush(...msgs: Item[]) {
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent))
-    for (const msg of msgs) {
-      queue.push(msg)
-    }
-    queue.flush()
-    return sent
-  }
-
-  it("merges repeated text deltas", () => {
-    const sent = items(flush(update("a", "a"), update("b", "b")))
-    expect(sent).toHaveLength(1)
-    expect(sent[0]!.part.text).toBe("ab")
-    expect(sent[0]!.delta?.textDelta).toBe("ab")
-  })
-
-  it("uses part messageID when messageID is blank", () => {
-    const sent = items(flush({ ...update("a", "a"), messageID: "" }, { ...update("b", "b"), messageID: "" }))
-    expect(sent).toHaveLength(1)
-    expect(sent[0]!.part.text).toBe("ab")
-  })
-
-  it("appends deltas onto queued full parts", () => {
-    const sent = items(flush(update("hello"), update(" world", " world")))
-    expect(sent[0]!.part.text).toBe("hello world")
-    expect(sent[0]!.delta).toBeUndefined()
-  })
-
-  it("keeps the latest full part", () => {
-    const sent = items(flush(update("old"), update("new")))
-    expect(sent[0]!.part.text).toBe("new")
-  })
-
-  it("flushes deltas before later full updates", () => {
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent))
-    queue.push(update("a", "a"))
-    queue.push(update("done"))
-    queue.flush()
-    const flat = items(sent)
-    expect(flat).toHaveLength(2)
-    expect(flat[0]!.part.text).toBe("a")
-    expect(flat[1]!.part.text).toBe("done")
-  })
-
-  it("batches multiple sessions into one message", () => {
-    const sent = flush(update("a", "a", "sess-1"), update("b", "b", "sess-2"))
-    expect(sent).toHaveLength(1)
-    expect(sent[0]!.type).toBe("partsUpdated")
-    expect(items(sent)).toHaveLength(2)
-  })
-
-  it("flushes focused sessions immediately", () => {
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent))
-    queue.push(update("a", "a", "sess-2"))
-    queue.focus("sess-2")
-    expect(items(sent)[0]!.part.text).toBe("a")
-  })
-
-  it("drop() discards queued updates and clears active", () => {
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent))
-    queue.focus("sess-1")
-    queue.push(update("a", "a", "sess-2"))
-    queue.drop("sess-2")
-    queue.drop("sess-1")
-    queue.flush()
-    expect(sent).toHaveLength(0)
-  })
-
-  it("never merges across sessions, messages, or parts", () => {
-    // Identity invariants: model-provider caching and session isolation depend on
-    // sessionID / messageID / partID never being swapped or combined by the scheduler.
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent))
-    queue.push({
-      ...update("a", "a", "sess-1"),
-      messageID: "m1",
-      part: { id: "p1", type: "text", messageID: "m1", text: "a" },
-    })
-    queue.push({
-      ...update("b", "b", "sess-2"),
-      messageID: "m1",
-      part: { id: "p1", type: "text", messageID: "m1", text: "b" },
-    })
-    queue.push({
-      ...update("c", "c", "sess-1"),
-      messageID: "m2",
-      part: { id: "p1", type: "text", messageID: "m2", text: "c" },
-    })
-    queue.push({
-      ...update("d", "d", "sess-1"),
-      messageID: "m1",
-      part: { id: "p2", type: "text", messageID: "m1", text: "d" },
-    })
-    queue.flush()
-    const flat = items(sent)
-    expect(flat).toHaveLength(4)
-    // Each tuple preserved exactly
-    const keys = flat.map((msg) => `${msg.sessionID}:${msg.messageID}:${(msg.part as { id: string }).id}`)
-    expect(new Set(keys).size).toBe(4)
-    expect(keys).toEqual(expect.arrayContaining(["sess-1:m1:p1", "sess-2:m1:p1", "sess-1:m2:p1", "sess-1:m1:p2"]))
-  })
-
-  it("preserves part fields and concatenates deltas in arrival order", () => {
-    // Full-part replacement must keep all incoming fields verbatim; text deltas must
-    // concatenate in arrival order so cumulative text matches what the CLI streamed.
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent))
-    queue.push({
-      type: "partUpdated",
-      sessionID: "sess-1",
-      messageID: "m1",
-      part: { id: "p1", type: "text", messageID: "m1", text: "hi", extraMetadata: { tokens: 42 } } as unknown as never,
-    })
-    queue.push({
-      type: "partUpdated",
-      sessionID: "sess-1",
-      messageID: "m1",
-      part: { id: "p1", type: "text", messageID: "m1", text: "hi world" } as unknown as never,
-      delta: { type: "text-delta", textDelta: " world" },
-    })
-    queue.push({
-      type: "partUpdated",
-      sessionID: "sess-1",
-      messageID: "m1",
-      part: { id: "p1", type: "text", messageID: "m1", text: "hi world!" } as unknown as never,
-      delta: { type: "text-delta", textDelta: "!" },
-    })
-    queue.flush()
-    const flat = items(sent)
-    expect(flat).toHaveLength(1)
-    const part = flat[0]!.part as { text: string; extraMetadata?: { tokens: number } }
-    // Initial full-part fields preserved
-    expect(part.extraMetadata?.tokens).toBe(42)
-    // Deltas concatenated in order onto the full part
-    expect(part.text).toBe("hi world!")
-    // When deltas fold into a queued full part, emit as a full-part replacement
-    // (no delta) so the webview does an authoritative replace rather than
-    // appending to whatever state it currently holds.
-    expect(flat[0]!.delta).toBeUndefined()
-  })
-
-  it("tracks stats across coalesced and batched emissions", () => {
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent))
-    queue.push(update("a", "a", "sess-1"))
-    queue.push(update("b", "b", "sess-1"))
-    queue.push(update("c", "c", "sess-2"))
-    queue.flush()
-    const stats = queue.stats()
-    expect(stats.received).toBe(3)
-    // 2 deltas to sess-1 merged into 1 update; sess-2 has 1 update; both emitted in one batch
-    expect(stats.emitted).toBe(2)
-    expect(stats.batches).toBe(1)
-  })
-
-  function sleep(ms: number) {
-    return new Promise<void>((r) => setTimeout(r, ms))
-  }
-
-  it("active session flushes on its cadence", async () => {
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent), {
-      activeMs: 5,
-      backgroundBaseMs: 1000,
-      backgroundStepMs: 0,
-      backgroundMaxMs: 1000,
-    })
-    queue.focus("sess-1")
-    queue.push(update("a", "a", "sess-1"))
-    queue.push(update("b", "b", "sess-1"))
-    await sleep(25)
-    expect(items(sent)[0]!.part.text).toBe("ab")
-    expect(queue.stats().active).toBe(1)
-    expect(queue.stats().background).toBe(0)
-    queue.dispose()
-  })
-
-  it("background throttles independent of active", async () => {
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent), {
-      activeMs: 5,
-      backgroundBaseMs: 20,
-      backgroundStepMs: 0,
-      backgroundMaxMs: 20,
-    })
-    queue.focus("sess-1")
-    queue.push(update("a", "a", "sess-1"))
-    queue.push(update("x", "x", "sess-2"))
-    queue.push(update("y", "y", "sess-3"))
-    await sleep(15)
-    // Active lane flushed alone first
-    expect(sent).toHaveLength(1)
-    await sleep(20)
-    // Background batch flushed after its own interval
-    expect(sent).toHaveLength(2)
-    expect(queue.stats().active).toBe(1)
-    expect(queue.stats().background).toBe(1)
-    queue.dispose()
-  })
-
-  it("focus(A→B) flushes B immediately and schedules A on background", async () => {
-    const sent: Sent[] = []
-    const queue = new SessionStreamScheduler((msg) => sent.push(msg as Sent), {
-      activeMs: 50,
-      backgroundBaseMs: 10,
-      backgroundStepMs: 0,
-      backgroundMaxMs: 10,
-    })
-    queue.focus("sess-1")
-    queue.push(update("a", "a", "sess-1"))
-    queue.push(update("b", "b", "sess-2"))
-    queue.focus("sess-2")
-    // Switching flushed sess-2's pending update immediately
-    expect(items(sent).some((item) => item.part.text === "b")).toBe(true)
-    await sleep(25)
-    // Sess-1 drained on background lane after focus switch
-    expect(items(sent).some((item) => item.part.text === "a")).toBe(true)
-    queue.dispose()
   })
 })
